backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-ollama
langchain-qdrant
qdrant-client
pypdf
python-docx
sentence-transformers
ollama
python-dotenv
reportlab
sqlalchemy
psycopg2-binary
supabase
werkzeug
python-jose[cryptography]
passlib[bcrypt]